// src/data/enums/specialisation.enum.ts
export enum Specialisation {
  AntiStigmatism = "Anti-Stigmatism",
  YouthMentalWellness = "Youth Mental Wellness",
  OCD = "Obsessive Compulsive Disorder",
  OverallMentalWellbeing = "Overall Mental Wellbeing",
  Therapy = "Therapy",
  Anxiety = "Anxiety",
  PostPartum = "Post-Partum",
  Corporate = "Corporate",
}